FinSec

+++ CSHT-Client/src/team/csht/entity/Good.java
@@ -11,20 +11,20 @@ public class Good implements Serializable {
     private String name;
     private String merchant;
     private String buyer;
-    private double price;
+    private float price;
     private boolean existence = true;
     private Icon icon;
 
     public Good() {
         super();
     }
-    public Good(int id, String name, double price) {
+    public Good(int id, String name, float price) {
         super();
         this.id=id;
         this.name=name;
         this.price=price;
     }
-    public Good(String name, double price){
+    public Good(String name, float price){
         super();
         this.name= name;
         this.price=price;
@@ -54,7 +54,7 @@ public class Good implements Serializable {
     public void setBuyer(String buyer) {
         this.buyer = buyer;
     }
-    public double getPrice() {
+    public float getPrice() {
         return price;
     }
     public void setPrice(float price) {
